TransfersDetailsModal, TransfersStatusModal updated, transfer helpers & components added
