+++ src/features/transfer/TransfersDetailsModal.tsx
@@ -1,50 +1,32 @@
 import Image from 'next/image';
-import { useEffect, useState } from 'react';
 
 import { MessageStatus, MessageTimeline, useMessageTimeline } from '@hyperlane-xyz/widgets';
 
-import { Spinner } from '../../components/animation/Spinner';
 import { ChainLogo } from '../../components/icons/ChainLogo';
 import { Modal } from '../../components/layout/Modal';
-import { links } from '../../consts/links';
 import ArrowRightIcon from '../../images/icons/arrow-right.svg';
-import CheckmarkCircleIcon from '../../images/icons/checkmark-circle.svg';
-import EnvelopeHeartIcon from '../../images/icons/envelope-heart.svg';
-import ErrorCircleIcon from '../../images/icons/error-circle.svg';
 import LinkIcon from '../../images/icons/external-link-icon.svg';
-import { toBase64 } from '../../utils/base64';
 import { formatTimestamp } from '../../utils/date';
-import { parseCaip2Id } from '../chains/caip2';
-import {
-  getChainDisplayName,
-  hasPermissionlessChain,
-  isPermissionlessChain,
-} from '../chains/utils';
-import { getMultiProvider } from '../multiProvider';
-import { useStore } from '../store';
+import { getChainDisplayName, hasPermissionlessChain } from '../chains/utils';
 import { getAllTokens } from '../tokens/metadata';
 import { isNativeToken } from '../tokens/native';
 import { useAccountForChain } from '../wallet/hooks';
 
+import { BasicSpinner } from './components/BasicSpinner';
+import { getHypExplorerLink, getTransferStatusLabel } from './helpers';
 import { TransferContext, TransferStatus } from './types';
 
 export function TransfersDetailsModal({
   isOpen,
   close,
-  transfers,
+  transfer,
 }: {
   isOpen: boolean;
   close: () => void;
-  transfers: TransferContext[];
+  transfer: TransferContext;
 }) {
-  const [index, setIndex] = useState(0);
-  // Auto update index to newest transfer when opening
-  useEffect(() => {
-    if (isOpen) setIndex(transfers.length - 1);
-  }, [isOpen, transfers.length]);
-
   const { params, status, originTxHash, msgId, timestamp, activeAccountAddress, route } =
-    transfers[index] || {};
+    transfer || {};
   const { destinationCaip2Id, originCaip2Id, tokenAddress, amount } = params || {};
 
   const account = useAccountForChain(originCaip2Id);
@@ -54,34 +36,13 @@ export function TransfersDetailsModal({
 
   const isPermissionlessRoute = hasPermissionlessChain([destinationCaip2Id, originCaip2Id]);
 
-  let statusDescription = '...';
-  if (!isAccountReady) statusDescription = 'Please connect wallet to continue';
-  else if (status === TransferStatus.Preparing)
-    statusDescription = 'Preparing for token transfer...';
-  else if (status === TransferStatus.CreatingApprove)
-    statusDescription = 'Preparing approve transaction...';
-  else if (status === TransferStatus.SigningApprove)
-    statusDescription = `Sign approve transaction in ${connectorName} to continue.`;
-  else if (status === TransferStatus.ConfirmingApprove)
-    statusDescription = 'Confirming approve transaction...';
-  else if (status === TransferStatus.CreatingTransfer)
-    statusDescription = 'Preparing transfer transaction...';
-  else if (status === TransferStatus.SigningTransfer)
-    statusDescription = `Sign transfer transaction in ${connectorName} to continue.`;
-  else if (status === TransferStatus.ConfirmingTransfer)
-    statusDescription = 'Confirming transfer transaction...';
-  else if (status === TransferStatus.ConfirmedTransfer)
-    if (!isPermissionlessRoute)
-      statusDescription = 'Transfer transaction confirmed, delivering message...';
-    else
-      statusDescription =
-        'Transfer confirmed, the funds will arrive when the message is delivered.';
-  else if (status === TransferStatus.Delivered)
-    statusDescription = 'Delivery complete, transfer successful!';
-  else if (status === TransferStatus.Failed)
-    statusDescription = 'Transfer failed, please try again.';
-
   const explorerLink = getHypExplorerLink(originCaip2Id, msgId);
+  const statusDescription = getTransferStatusLabel(
+    status,
+    connectorName,
+    isPermissionlessRoute,
+    isAccountReady,
+  );
   const date = timestamp ? formatTimestamp(timestamp) : formatTimestamp(new Date().getTime());
   const token = getAllTokens().find((t) => t.address === tokenAddress);
 
@@ -119,7 +80,7 @@ export function TransfersDetailsModal({
         {isPermissionlessRoute ? (
           <BasicSpinner transferStatus={status} />
         ) : (
-          <Timeline transferStatus={status} transferIndex={index} originTxHash={originTxHash} />
+          <Timeline transferStatus={status} originTxHash={originTxHash} />
         )}
         {status !== TransferStatus.ConfirmedTransfer && status !== TransferStatus.Delivered ? (
           <div
@@ -250,11 +211,9 @@ export function TransfersDetailsModal({
 
 function Timeline({
   transferStatus,
-  transferIndex,
   originTxHash,
 }: {
   transferStatus: TransferStatus;
-  transferIndex: number;
   originTxHash?: string;
 }) {
   const isFailed = transferStatus === TransferStatus.Failed;
@@ -262,14 +221,6 @@ function Timeline({
     originTxHash: isFailed ? undefined : originTxHash,
   });
   const messageStatus = isFailed ? MessageStatus.Failing : message?.status || MessageStatus.Pending;
-
-  // TODO move this so it runs even if timeline isn't open
-  const updateTransferStatus = useStore((s) => s.updateTransferStatus);
-  useEffect(() => {
-    if (messageStatus === MessageStatus.Delivered) {
-      updateTransferStatus(transferIndex, TransferStatus.Delivered);
-    }
-  }, [transferIndex, messageStatus, updateTransferStatus]);
 
   return (
     <div className="mt-4 mb-7 w-full flex flex-col justify-center items-center timeline-container">
@@ -282,52 +233,4 @@ function Timeline({
       />
     </div>
   );
-}
-
-function BasicSpinner({ transferStatus }: { transferStatus: TransferStatus }) {
-  let content;
-  if (transferStatus === TransferStatus.Delivered) {
-    content = (
-      <Image
-        src={CheckmarkCircleIcon}
-        alt="Delivered"
-        width={80}
-        height={80}
-        className="opacity-80"
-      />
-    );
-  } else if (transferStatus === TransferStatus.ConfirmedTransfer) {
-    content = (
-      <Image
-        src={EnvelopeHeartIcon}
-        alt="Delivering"
-        width={80}
-        height={80}
-        className="opacity-80"
-      />
-    );
-  } else if (transferStatus === TransferStatus.Failed) {
-    content = (
-      <Image src={ErrorCircleIcon} alt="Failed" width={86} height={86} className="opacity-80" />
-    );
-  } else {
-    content = <Spinner />;
-  }
-  return <div className="py-4 flex flex-col justify-center items-center">{content}</div>;
-}
-
-// TODO test with solana chain config, or disallow it
-function getHypExplorerLink(originCaip2Id: Caip2Id, msgId?: string) {
-  if (!originCaip2Id || !msgId) return null;
-  const baseLink = `${links.explorer}/message/${msgId}`;
-  if (isPermissionlessChain(originCaip2Id)) {
-    const { reference } = parseCaip2Id(originCaip2Id);
-    const chainConfig = getMultiProvider().getChainMetadata(reference);
-    const serializedConfig = toBase64([chainConfig]);
-    if (serializedConfig) {
-      const params = new URLSearchParams({ chains: serializedConfig });
-      return `${baseLink}?${params.toString()}`;
-    }
-  }
-  return baseLink;
 }
