+++ src/features/transfer/TransfersStatusModal.tsx
@@ -1,23 +1,16 @@
-import Image from 'next/image';
 import { useEffect, useState } from 'react';
 
 import { MessageStatus, MessageTimeline, useMessageTimeline } from '@hyperlane-xyz/widgets';
 
-import { Spinner } from '../../components/animation/Spinner';
 import { IconButton } from '../../components/buttons/IconButton';
 import { ChevronIcon } from '../../components/icons/Chevron';
 import { Modal } from '../../components/layout/Modal';
-import { links } from '../../consts/links';
-import CheckmarkCircleIcon from '../../images/icons/checkmark-circle.svg';
-import EnvelopeHeartIcon from '../../images/icons/envelope-heart.svg';
-import ErrorCircleIcon from '../../images/icons/error-circle.svg';
-import { toBase64 } from '../../utils/base64';
-import { parseCaip2Id } from '../chains/caip2';
-import { hasPermissionlessChain, isPermissionlessChain } from '../chains/utils';
-import { getMultiProvider } from '../multiProvider';
+import { hasPermissionlessChain } from '../chains/utils';
 import { useStore } from '../store';
 import { useAccountForChain } from '../wallet/hooks';
 
+import { BasicSpinner } from './components/BasicSpinner';
+import { getHypExplorerLink, getTransferStatusLabel } from './helpers';
 import { TransferContext, TransferStatus } from './types';
 
 export function TransfersStatusModal({
@@ -45,32 +38,12 @@ export function TransfersStatusModal({
 
   const isPermissionlessRoute = hasPermissionlessChain([destinationCaip2Id, originCaip2Id]);
 
-  let statusDescription = '...';
-  if (!isAccountReady) statusDescription = 'Please connect wallet to continue';
-  else if (status === TransferStatus.Preparing)
-    statusDescription = 'Preparing for token transfer...';
-  else if (status === TransferStatus.CreatingApprove)
-    statusDescription = 'Preparing approve transaction...';
-  else if (status === TransferStatus.SigningApprove)
-    statusDescription = `Sign approve transaction in ${connectorName} to continue.`;
-  else if (status === TransferStatus.ConfirmingApprove)
-    statusDescription = 'Confirming approve transaction...';
-  else if (status === TransferStatus.CreatingTransfer)
-    statusDescription = 'Preparing transfer transaction...';
-  else if (status === TransferStatus.SigningTransfer)
-    statusDescription = `Sign transfer transaction in ${connectorName} to continue.`;
-  else if (status === TransferStatus.ConfirmingTransfer)
-    statusDescription = 'Confirming transfer transaction...';
-  else if (status === TransferStatus.ConfirmedTransfer)
-    if (!isPermissionlessRoute)
-      statusDescription = 'Transfer transaction confirmed, delivering message...';
-    else
-      statusDescription =
-        'Transfer confirmed, the funds will arrive when the message is delivered.';
-  else if (status === TransferStatus.Delivered)
-    statusDescription = 'Delivery complete, transfer successful!';
-  else if (status === TransferStatus.Failed)
-    statusDescription = 'Transfer failed, please try again.';
+  const statusDescription = getTransferStatusLabel(
+    status,
+    connectorName,
+    isPermissionlessRoute,
+    isAccountReady,
+  );
 
   const explorerLink = getHypExplorerLink(originCaip2Id, msgId);
 
@@ -158,52 +131,4 @@ function Timeline({
       />
     </div>
   );
-}
-
-function BasicSpinner({ transferStatus }: { transferStatus: TransferStatus }) {
-  let content;
-  if (transferStatus === TransferStatus.Delivered) {
-    content = (
-      <Image
-        src={CheckmarkCircleIcon}
-        alt="Delivered"
-        width={80}
-        height={80}
-        className="opacity-80"
-      />
-    );
-  } else if (transferStatus === TransferStatus.ConfirmedTransfer) {
-    content = (
-      <Image
-        src={EnvelopeHeartIcon}
-        alt="Delivering"
-        width={80}
-        height={80}
-        className="opacity-80"
-      />
-    );
-  } else if (transferStatus === TransferStatus.Failed) {
-    content = (
-      <Image src={ErrorCircleIcon} alt="Failed" width={86} height={86} className="opacity-80" />
-    );
-  } else {
-    content = <Spinner />;
-  }
-  return <div className="py-4 flex flex-col justify-center items-center">{content}</div>;
-}
-
-// TODO test with solana chain config, or disallow it
-function getHypExplorerLink(originCaip2Id: Caip2Id, msgId?: string) {
-  if (!originCaip2Id || !msgId) return null;
-  const baseLink = `${links.explorer}/message/${msgId}`;
-  if (isPermissionlessChain(originCaip2Id)) {
-    const { reference } = parseCaip2Id(originCaip2Id);
-    const chainConfig = getMultiProvider().getChainMetadata(reference);
-    const serializedConfig = toBase64([chainConfig]);
-    if (serializedConfig) {
-      const params = new URLSearchParams({ chains: serializedConfig });
-      return `${baseLink}?${params.toString()}`;
-    }
-  }
-  return baseLink;
 }
