+++ src/features/transfer/helpers.ts
@@ -1,0 +1,59 @@
+import { links } from '../../consts/links';
+import { toBase64 } from '../../utils/base64';
+import { parseCaip2Id } from '../chains/caip2';
+import { isPermissionlessChain } from '../chains/utils';
+import { getMultiProvider } from '../multiProvider';
+
+import { TransferStatus } from './types';
+
+// TODO test with solana chain config, or disallow it
+export function getHypExplorerLink(originCaip2Id: Caip2Id, msgId?: string) {
+  if (!originCaip2Id || !msgId) return null;
+  const baseLink = `${links.explorer}/message/${msgId}`;
+  if (isPermissionlessChain(originCaip2Id)) {
+    const { reference } = parseCaip2Id(originCaip2Id);
+    const chainConfig = getMultiProvider().getChainMetadata(reference);
+    const serializedConfig = toBase64([chainConfig]);
+    if (serializedConfig) {
+      const params = new URLSearchParams({ chains: serializedConfig });
+      return `${baseLink}?${params.toString()}`;
+    }
+  }
+  return baseLink;
+}
+
+export function getTransferStatusLabel(
+  status: TransferStatus,
+  connectorName: string,
+  isPermissionlessRoute: boolean,
+  isAccountReady: boolean,
+) {
+  let statusDescription = '...';
+  if (!isAccountReady) statusDescription = 'Please connect wallet to continue';
+  else if (status === TransferStatus.Preparing)
+    statusDescription = 'Preparing for token transfer...';
+  else if (status === TransferStatus.CreatingApprove)
+    statusDescription = 'Preparing approve transaction...';
+  else if (status === TransferStatus.SigningApprove)
+    statusDescription = `Sign approve transaction in ${connectorName} to continue.`;
+  else if (status === TransferStatus.ConfirmingApprove)
+    statusDescription = 'Confirming approve transaction...';
+  else if (status === TransferStatus.CreatingTransfer)
+    statusDescription = 'Preparing transfer transaction...';
+  else if (status === TransferStatus.SigningTransfer)
+    statusDescription = `Sign transfer transaction in ${connectorName} to continue.`;
+  else if (status === TransferStatus.ConfirmingTransfer)
+    statusDescription = 'Confirming transfer transaction...';
+  else if (status === TransferStatus.ConfirmedTransfer)
+    if (!isPermissionlessRoute)
+      statusDescription = 'Transfer transaction confirmed, delivering message...';
+    else
+      statusDescription =
+        'Transfer confirmed, the funds will arrive when the message is delivered.';
+  else if (status === TransferStatus.Delivered)
+    statusDescription = 'Delivery complete, transfer successful!';
+  else if (status === TransferStatus.Failed)
+    statusDescription = 'Transfer failed, please try again.';
+
+  return statusDescription;
+}
